+++ react/__tests__/components/ProductSpecifications.test.js
@@ -4,8 +4,18 @@ import ProductSpecifications from '../../ProductSpecifications'
 import { render } from '@vtex/test-tools/react'
 
 describe('<ProductSpecifications /> component', () => {
+  const defaultProps = {
+    specificationsDefault: {
+      showSpecifications: true,
+    },
+  }
+
   const renderComponent = customProps => {
-    const comp = <ProductSpecifications {...customProps} />
+    const props = {
+      ...defaultProps,
+      ...customProps,
+    }
+    const comp = <ProductSpecifications {...props} />
 
     return render(comp)
   }
